Break out table components into separate files
